+++ lib/kapnismology/rake_task.rb
@@ -1,12 +1,19 @@
+require 'kapnismology/terminal'
 module Kapnismology
   class RakeTask
     def output
       evaluations = SmokeTest.evaluations
+      puts
+      puts
       evaluations.each do |evaluation|
         puts evaluation.to_s
         puts
       end
-      raise 'We have some failures in our smoke tests' unless evaluations.passed?
+      if evaluations.passed?
+        puts Terminal.green('All smoke tests passed successfully!')
+      else
+        abort Terminal.red('We have some failures in our smoke tests')
+      end
     end
   end
 end